[cheriabi] Use GType gobject type ids
Use of `gobject` type ids in gtk must be updated to use `GType` (which
is a capability width/aligned type). When type ids are initialised the
Use of `g_once_init_enter/leave` must also be updated to the equivalent
functions for pointer width values `g_once_enter/leave_pointer`
(available in glib 2.80 and backported to glib 2.76 on CheriBSD).